# Conflicts:
#	services/src/test/java/ar/edu/itba/paw/services/WeeklyDigestServiceImplTest.java

+++ services/src/test/java/ar/edu/itba/paw/services/CommunityServiceImplTest.java
@@ -1139,7 +1139,7 @@ class CommunityServiceImplTest {
 
         @Override
         public void sendWeeklyModeratorDigest(final List<EmailRecipient> moderatorRecipients,
-                                              final int pendingRequestCount) {
+                                              final long pendingRequestCount) {
         }
 
         @Override
